test(onboard): consolidate trace handoff proof
